Add tempo (100)
Note that voice isolation is Alt-2
A simple L (l) over the note brings up the line symbols (molto accelarando)
And instead of 32 notes, visually more appealing to rewrite as 16 with accelerando (or molto)
thrown in on top. To extend the line just drag from bottom right.

And test the chords by playing it out. You can write a simpler version that has the
same effect.